+++ src/services/fatecAPI.ts
@@ -1,9 +1,6 @@
-// Connection with fatec api
-
-// const fatecApi = require('fatec-api')
 import fatecApi from 'fatec-api'
 
-const apiFatec = (login, password) => {
+const apiFatec = (login: string, password: string) => {
   const myAccount = new fatecApi.Account(login, password)
   return myAccount