+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-03-28
+
+### Changed 
+
+- Refactoring the code to TS;
+- Removed sequelize;
+
+### Added 
+
+- Added typeorm;
+- Added migration and models of db;
+
+### Changed 
+
+- Login funcion verify if RA student exist in db;
+
 ## [0.1.0] - 2021-03-23
 
 ### Added